remove unused gethitssize, revert getmemsize to size_t

+++ Buffer_shared.cpp
@@ -78,7 +78,7 @@ void FacetHistogramBuffer::Reset() {
 }
 
 size_t FacetHistogramBuffer::GetMemSize() const {
-	int sum = 0;
+	size_t sum = 0;
 	sum += nbHitsHistogram.capacity() * sizeof(double);
 	sum += distanceHistogram.capacity() * sizeof(double);
 #ifdef MOLFLOW
